Add translate scripts
