+++ src/small_thoughts/__init__.py
@@ -3,11 +3,11 @@ from .utils.deduplicate import deduplicate
 from .utils.reason import reason
 from .utils.mix import mix_subsets
 from .utils.eval import EVALUATION_DATASETS
-from .utils.prompt import SYSTEM_PROMPT, MIX_PROMPT
-from .science import science_main
-from .puzzle import puzzle_main
-from .maths import math_main
-from .code import code_main
+from .utils.prompt import REASONING_PROMPT, TRANSLATION_PROMPT
+from .reasoning.science import science_main
+from .reasoning.puzzle import puzzle_main
+from .reasoning.maths import math_main
+from .reasoning.code import code_main
 from .generation import main
 
-__all__ = ["deduplicate", "decontaminate", "reason", "mix_subsets", "EVALUATION_DATASETS", "SYSTEM_PROMPT", "MIX_PROMPT", "science_main", "puzzle_main", "math_main", "code_main", "main"]
+__all__ = ["deduplicate", "decontaminate", "reason", "mix_subsets", "EVALUATION_DATASETS", "REASONING_PROMPT", "TRANSLATION_PROMPT", "science_main", "puzzle_main", "math_main", "code_main", "main"]
