+++ src/small_thoughts/reasoning/maths/__init__.py
