+++ src/small_thoughts/reasoning/puzzle/__init__.py
