+++ src/small_thoughts/translation/smoltalk/__init__.py
@@ -1,0 +1,1 @@
+from ..smoltalk.smoltalk import main as smoltalk_main
