+++ src/small_thoughts/utils/reason.py
@@ -1,6 +1,6 @@
 from bespokelabs import curator
 
-from ..utils.prompt import SYSTEM_PROMPT
+from ..utils.prompt import REASONING_PROMPT
 
 
 class Reasoner(curator.LLM):
@@ -35,6 +35,6 @@ def reason(dataset, base_url, model_name, temperture=0.0, max_tokens=8192, syste
         model_name=model_name,
         generation_params={"temp": temperture, "max_tokens": max_tokens},
         backend_params={"base_url": base_url, "max_requests_per_minute": max_requests_per_minute, "max_tokens_per_minute": max_tokens_per_minute},
-        system_prompt=SYSTEM_PROMPT[system_prompt_type],
+        system_prompt=REASONING_PROMPT[system_prompt_type],
     )
     return reasoner(dataset)
