+++ src/small_thoughts/utils/translate.py
@@ -1,0 +1,36 @@
+from bespokelabs import curator
+
+from ..utils.prompt import TRANSLATION_PROMPT
+
+
+class Translater(curator.LLM):
+    return_completions_object = True
+
+    def __init__(self, model_name, generation_params, backend_params, system_prompt=None):
+        super().__init__(model_name=model_name, generation_params=generation_params, backend_params=backend_params, backend="openai")
+        self.system_prompt = system_prompt
+
+    def prompt(self, input):
+        """Create a prompt for the LLM to translate about the content."""
+        content = input["content"]
+        return [
+            {"role": "system", "content": self.system_prompt},
+            {"role": "user", "content": content},
+        ]
+
+    def parse(self, input, response):
+        """Parse the LLM response to extract translated content."""
+        return {
+            "content": input["content"],
+            "translated_content": response["choices"][0]["message"]["content"],
+        }
+
+
+def translate(dataset, base_url, model_name, temperture=0.0, max_tokens=8192, system_prompt_type="english", max_requests_per_minute=1_000, max_tokens_per_minute=1_000_000_000):
+    reasoner = Translater(
+        model_name=model_name,
+        generation_params={"temp": temperture, "max_tokens": max_tokens},
+        backend_params={"base_url": base_url, "max_requests_per_minute": max_requests_per_minute, "max_tokens_per_minute": max_tokens_per_minute},
+        system_prompt=TRANSLATION_PROMPT[system_prompt_type],
+    )
+    return reasoner(dataset)
